requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
boto==2.49.0
boto3==1.17.64
botocore==1.20.64
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
dj-database-url==0.5.0
django-bootstrap3==14.2.0
django-braces==1.14.0
django-debug-toolbar==3.2
django-heroku==0.3.1
django-storages==1.11.1
django-widget-tweaks==1.4.8
Faker==5.8.0
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
misaka==2.1.1
Pillow==8.1.0
psycopg2-binary==2.8.6
pycparser==2.20
python-dateutil==2.8.1
python-decouple==3.4
requests==2.25.1
s3transfer==0.4.2
six==1.15.0
stripe==2.56.0
text-unidecode==1.3
urllib3==1.26.3
whitenoise==5.2.0
wincertstore==0.2
